requirements.txt
requests==2.32.3
pyspark==3.5.5
apache-airflow==2.8.1